requirements.txt
fastapi==0.115.12
sqlalchemy==2.0.23
pyjwt==2.8.0
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
uvicorn==0.34.0
pydantic==2.11.7
email-validator==2.1.0.post1
pydantic-settings>=2.0